hackathon/requirements.txt
opencv-python-headless
facenet-pytorch
hsemotion